requirements.txt
streamlit
requests
trafilatura
pypdf
python-dotenv
serpapi
groq